RR-231: Move event form to client Prepare for Details to be moved
