+++ src/client/modules/Events/EventForm/index.jsx
@@ -5,9 +5,9 @@ import GridRow from '@govuk-react/grid-row'
 import GridCol from '@govuk-react/grid-col'
 
 import urls from '../../../../lib/urls'
-import LocalHeader from '../../../../client/components/LocalHeader/LocalHeader.jsx'
-import { Main } from '../../../../client/components'
-import TaskForm from '../../../../client/components/Task/Form'
+import LocalHeader from '../../../components/LocalHeader/LocalHeader.jsx'
+import { Main } from '../../../components'
+import TaskForm from '../../../components/Task/Form'
 import { TASK_GET_EVENTS_FORM_AND_METADATA, TASK_SAVE_EVENT } from './state'
 import { EventFormFields } from './EventFormFields'
 import { transformEventFormForAPIRequest } from './transformers'
